// src/app/models/sjcs/FundamentosCalificacionItem.ts
export class FundamentosCalificacionItem {

  idFundamento: string;
  codigo: string;
  codigoEjis: string;
  idInstitucion: string
  descripcionFundamento: string;
  fechabaja: string;
  idTipoDictamenEjg: string;
  descripcionDictamen: string;
  textoEnPlantilla: string;
  idDescripcion: string


  constructor() { }
}